fix: cleanup and moving functions to helper.js

+++ backend/controllers/users-controller.js
@@ -1,10 +1,9 @@
 const { validationResult } = require('express-validator');
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
-const tj = require('@mapbox/togeojson');
-const fs = require('fs');
-const DOMParser = require('xmldom').DOMParser;
 const mongoose = require('mongoose');
+const Helper = require('./helper.js');
+
 
 const User = require('../models/user.js');
 const Route = require('../models/route.js');
@@ -59,6 +58,7 @@ const login = async(req, res, next) => {
     userID: recognizedUser.id,
     username: recognizedUser.username,
     email: recognizedUser.email,
+    polarAffiliated: recognizedUser.polarAffiliated,
     token: token
   })
 }
@@ -150,6 +150,7 @@ const createUser = async(req, res, next) => {
     userID: newUser.id,
     username: newUser.username,
     email: newUser.email,
+    polarAffiliated: newUser.polarAffiliated,
     token: token
   });
 }
@@ -232,6 +233,7 @@ const connectToPolarAPI = async(req, res, next) => {
         const sess = await mongoose.startSession();
         sess.startTransaction();
         user.polarID = apiID;
+        user.polarAffiliated = true;
         await user.save({ session: sess });
         await sess.commitTransaction();
       }catch(err) {
@@ -255,7 +257,7 @@ const connectToPolarAPI = async(req, res, next) => {
     }); 
 
     // proper user check needed later 
-    const responseData2 = await data2.json();
+    const responseData2 = await data2.text();
     console.log(responseData2);  
   }catch(err) {
     throw new Error(err);
@@ -325,7 +327,7 @@ const getNewData = async(req, res, next) => {
   }catch(err) {
     throw new Error(err);
   } 
-  const status = await filterExercise(uid, list, token);
+  const status = await Helper.filterExercise(uid, list, token);
   if (status.code !== 201) {
     throw status;
   }
@@ -355,9 +357,6 @@ const getUserRoutes = async(req, res, next) => {
     routeTypeList.push(type);
     type = iterator.next().value;
   }
-
-  
-
   res.status(200).json({ types: routeTypeList, list: user.routes });
 }
 
@@ -383,10 +382,6 @@ const getAllUserRoutes = async(req, res, next) => {
     keys.push(type);
     type = iterator.next().value;
   }
-
-
-
-  
 
   res.status(200).json({ keys: keys, values: user.routes });
 }
@@ -443,148 +438,6 @@ const getUserGeneralInfo = async(req, res, next) => {
   res.status(200).json({ info: userInfo });
 }
 
-const filterExercise = async(uid, exercises, token) => {
-  let userAuthorization = 'Bearer ' + token;
-
-  let status;
-  for (let i = 0; i < exercises.length; i++) {
-    try {
-      const data2 = await fetch(exercises[i], 
-      {
-        method: 'GET',
-        headers: {
-          'Accept': 'application/json',
-          'Authorization': userAuthorization
-        },
-      });
-  
-      // proper user check needed later 
-      const responseData = await data2.json();
-      if (responseData["has-route"]) {
-        const coord = await getPoints(exercises[i], token);
-        responseData['coord'] = coord;
-        status = await createRoute(uid, responseData, coord);
-
-        if (status.code !== 201) {
-          return status;
-        }
-      }
-    }catch(err) {
-      throw new Error(err);   
-    }    
-  }
-
-  return status;
-}
-
-const getPoints = async(exerciseURL, token) => {
-  let userAuthorization = 'Bearer ' + token;
-  let coord;
-
-  try {
-    const url = exerciseURL + '/gpx';
-    const data2 = await fetch(url, 
-    {
-      method: 'GET',
-      headers: {
-        'Accept': 'application/gpx+xml',
-        'Authorization': userAuthorization
-      },
-    });
-
-    // proper user check needed later 
-    // const responseData2 = await data2.text();
-    const gpxFile = "./newPoints.gpx"
-    const fileContent = await data2.text();
-    fs.writeFileSync(gpxFile, fileContent);
-    coord = await getRoute(gpxFile);
-    fs.unlink(gpxFile, (err) => {
-      if (err) {
-        console.error(err);
-      }
-    });    
-  }catch(err) {
-    throw new Error(err);   
-  }    
-  
-  return coord; 
-}
-
-const getRoute = async(file) => {
-  const gpx = new DOMParser().parseFromString(fs.readFileSync(file, 'utf8'));
-  const converted = tj.gpx(gpx);
-
-  const coord = converted.features[0].geometry.coordinates;
-
-  for (let i = 0; i < coord.length; i++) {
-    coord[i].pop();
-    [coord[i][0], coord[i][1]] = [coord[i][1], coord[i][0]];
-  }
-  
-  return coord;
-}
-
-const createRoute = async(uid, route, coord) => {
-  
-
-  // check if user exists
-  let user;
-  try {
-    user = await User.findById(uid);
-  } catch(err) {
-    return new HttpError('request error', 404);   
-  }
-
-  if (!user) {
-    return new HttpError('could not find user', 500);   
-  }
-
-  // create new route
-  const createdRoute = new Route({
-    user: user,
-    method: "POLAR",
-    type: route['detailed-sport-info'],
-    distance: route['distance'],
-    date: route['start-time'],
-    duration: route['duration'],
-    calories: route['calories'],
-    points: coord,
-    heartRate: route['heart-rate'],
-    // temp
-    other: route
-  });
-
-  // add new key if it doesn't exist in maps
-  if (user.routes.get(route['detailed-sport-info']) === undefined) {
-    user.routes.set(route['detailed-sport-info'], []);
-  }
-
-  try {
-    // transaction and sessions
-    // makes sure everything is good before putting both in the database
-    // manually create collection for posts
-    const sess = await mongoose.startSession();
-    sess.startTransaction();
-
-    await createdRoute.save( { session: sess });
-
-    user.routes.get(route['detailed-sport-info']).push(createdRoute._id);
-    await user.save({ session: sess });
-
-    await sess.commitTransaction();
-  } catch (err) {
-    console.log(err);
-    const error = new HttpError(
-      err, 500
-      //"Creation of route type Failed", 500
-    );
-    return error;
-  }
-
-  // temporary languuge for now
-  return new HttpError('Success', 201);
-}
-
 const deleteAccount = async(req, res, next) => {
   const uid = req.params.uid;
   const apiID = req.params.apiID;
@@ -605,23 +458,24 @@ const deleteAccount = async(req, res, next) => {
     throw new HttpError('user does not exist', 401);
   }
 
-  // delink polar account
-  try {
-    const response = async() => 
-    await fetch(`https://www.polaraccesslink.com/v3/users/${apiID}`, {
-      method: 'DELETE',
-      headers: {
-        'Authorization': userAuthorization
-      },
-    });
+  if (user.polarID !== 0) {
+    // delink polar account
+    try {
+      const response = async() => 
+      await fetch(`https://www.polaraccesslink.com/v3/users/${apiID}`, {
+        method: 'DELETE',
+        headers: {
+          'Authorization': userAuthorization
+        },
+      });
 
 
-    await response();
+      await response();
 
-  }catch(err) {
-    throw new HttpError(err, 500);
+    }catch(err) {
+      throw new HttpError(err, 500);
+    }    
   }
-
   // delete routes from routes
   // delete routes from routeType
   // delete user from db
@@ -634,12 +488,15 @@ const deleteAccount = async(req, res, next) => {
 
     // delete routes from routes
     const len = routes.length;
+
     for (let i = 0; i < len; i++) {
       await routes[i].deleteOne({session:sess});
     }
 
+    console.log("test");
     await user.deleteOne({ session: sess});
 
+    console.log("test2");
 
     await sess.commitTransaction();
   }catch(err) {
@@ -649,14 +506,11 @@ const deleteAccount = async(req, res, next) => {
   res.status(202).json({ message: "Successfully deleted account" });
 }
 
-
-
 exports.login = login;
 exports.createUser = createUser;
 exports.connectToPolarAPI = connectToPolarAPI;
 exports.getNewData = getNewData;
 exports.getUserRoutes = getUserRoutes;
-exports.deleteAccount = deleteAccount;
-
 exports.getAllUserRoutes = getAllUserRoutes;
 exports.getUserGeneralInfo = getUserGeneralInfo;
+exports.deleteAccount = deleteAccount;
